feat: add markdown formatting
