+++ src/pages/FormatMarkdown.tsx
@@ -1,25 +1,35 @@
 import { useCallback, useState } from "react";
-import { preprocessText, restoreText } from "../text-processing";
 import { CopyTextEntry } from "../components/copyTextEntry";
+import { preprocessText, restoreText } from "../text-processing";
 
-function removeMarkdownEmphasis(text: string): string {
-  return text.replace(/\*\*(.*?)\*\*/gm, "$1").replace(/_(.*?)_/gm, "$1");
+function replaceMarkdownElements(text: string) {
+  return text
+    .replace(/\n\* {1,}/g, "- ")
+    .replace(/( {4}|\t)\* {1,}/g, "    - ")
+    .replace(/( {4}|\t)/g, "   ")
+    .replace(/(\s*)(\d*\.) {2,}/gm, "$1$2 ");
 }
 
-export function EchoText() {
+export function FormatMarkdown() {
   const [count, setCount] = useState(1);
   const increase = useCallback(() => setCount((c) => c + 1), [setCount]);
 
   return (
     <>
-      <h1 className="mb-8 w-full text-center text-3xl font-bold">Echo Text</h1>
+      <h1 className="mb-8 w-full text-center text-3xl font-bold">
+        Format Markdown
+      </h1>
+
+      <h2 className="my-8 w-full text-center text-2xl font-bold">
+        Text Entries
+      </h2>
 
       <div className="my-5 flex flex-col gap-4">
         {Array.from(Array(count)).map((_, index) => (
           <CopyTextEntry
             key={index}
-            onBlur={(text) => removeMarkdownEmphasis(preprocessText(text))}
-            onCopy={restoreText}
+            onBlur={preprocessText}
+            onCopy={(text) => replaceMarkdownElements(restoreText(text))}
           />
         ))}
       </div>
